got rid of view in unit tests
