+++ metadata-selector/src/test/java/io/github/sergeivisotsky/metadata/selector/dao/impl/FormMetadataDaoImplTest.java
@@ -16,11 +16,10 @@
 
 package io.github.sergeivisotsky.metadata.selector.dao.impl;
 
-import java.util.List;
-
 import io.github.sergeivisotsky.metadata.selector.dao.AbstractMetadataDao;
-import io.github.sergeivisotsky.metadata.selector.dto.LookupHolder;
-import io.github.sergeivisotsky.metadata.selector.dto.LookupMetadata;
+import io.github.sergeivisotsky.metadata.selector.dto.form.FormField;
+import io.github.sergeivisotsky.metadata.selector.dto.form.FormMetadata;
+import io.github.sergeivisotsky.metadata.selector.dto.form.FormSection;
 import io.github.sergeivisotsky.metadata.selector.mapper.MetadataMapper;
 import org.junit.Before;
 import org.junit.Test;
@@ -37,23 +36,26 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
- * Unit test for {@link LookupMetadataDaoImpl}.
+ * Unit test for {@link FormMetadataDaoImpl}.
  *
  * @author Sergei Visotsky
  */
-public class LookupViewMetadataDaoImplTest extends AbstractMetadataDao {
+public class FormMetadataDaoImplTest extends AbstractMetadataDao {
 
     @Mock
     private NamedParameterJdbcTemplate jdbcTemplate;
 
     @Mock
-    private MetadataMapper<LookupHolder> lookupHolderMapper;
+    private MetadataMapper<FormMetadata> formMetadataMapper;
 
     @Mock
-    private MetadataMapper<LookupMetadata> lookupMetadataMapper;
+    private MetadataMapper<FormSection> formSectionMapper;
+
+    @Mock
+    private MetadataMapper<FormField> formFieldMapper;
 
     @InjectMocks
-    private LookupMetadataDaoImpl dao;
+    private FormMetadataDaoImpl dao;
 
     @Before
     public void setUp() {
@@ -61,19 +63,13 @@ public class LookupViewMetadataDaoImplTest extends AbstractMetadataDao {
     }
 
     @Test
-    public void shouldGetLookupMetadata() {
-        LookupHolder lookupHolder = new LookupHolder();
-        lookupHolder.setName("someHolder");
-        lookupHolder.setWeight(300);
-        lookupHolder.setHeight(20);
-        lookupHolder.setMetadata(List.of());
+    public void shouldGetFormMetadata() {
+        when(formMetadataMapper.getSql()).thenReturn("SELECT * FROM some_table WHERE id = 1");
+        when(formSectionMapper.getSql()).thenReturn("SELECT * FROM some_table WHERE id = 1");
+        when(formFieldMapper.getSql()).thenReturn("SELECT * FROM some_table WHERE id = 1");
 
-        when(lookupHolderMapper.getSql()).thenReturn("SELECT * FROM some_table WHERE id = 1");
-        when(jdbcTemplate.queryForObject(anyString(), anyMap(), any(RowMapper.class))).thenReturn(lookupHolder);
+        dao.getFormMetadata("someForm", "EN");
 
-        dao.getLookupMetadata("someLookup", "en");
-
-        verify(lookupHolderMapper).getSql();
         verify(jdbcTemplate).queryForObject(anyString(), anyMap(), any(RowMapper.class));
     }
 }
